new topics for the 1st time

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 size_t print_list(const list_t *h)
 {
 
-	int element = 0;
+	size_t element = 0;
 
 		while (h != NULL)
 		{
